+++ src/samples/permissionhooks/AllowlistPlugin.sol
@@ -68,25 +68,25 @@ contract AllowlistPlugin is IValidationHook, BasePlugin {
         selectorAllowlist[target][selector][msg.sender] = allowed;
     }
 
-    function preUserOpValidationHook(uint8 validationId, PackedUserOperation calldata userOp, bytes32)
+    function preUserOpValidationHook(uint32 validationId, PackedUserOperation calldata userOp, bytes32)
         external
         view
         override
         returns (uint256)
     {
-        if (validationId == uint8(ValidationId.PRE_VALIDATION_HOOK)) {
+        if (validationId == uint32(ValidationId.PRE_VALIDATION_HOOK)) {
             _checkAllowlistCalldata(userOp.callData);
             return 0;
         }
         revert NotImplemented();
     }
 
-    function preRuntimeValidationHook(uint8 validationId, address, uint256, bytes calldata data, bytes calldata)
+    function preRuntimeValidationHook(uint32 validationId, address, uint256, bytes calldata data, bytes calldata)
         external
         view
         override
     {
-        if (validationId == uint8(ValidationId.PRE_VALIDATION_HOOK)) {
+        if (validationId == uint32(ValidationId.PRE_VALIDATION_HOOK)) {
             _checkAllowlistCalldata(data);
             return;
         }
